TP5/requirements.txt
gymnasium[box2d]
stable-baselines3
Pillow
